requirements.txt
requests==2.28.2
beautifulsoup4==4.12.0
python-dotenv==1.0.0
pymongo~=4.3.3
flask~=2.2.3
gunicorn